parseFloat

+++ rxjs-1/race.js
@@ -15,10 +15,10 @@ class Race extends EventEmitter {
       }
     ).map(data => ({
       ...data,
-      'Time car 1': parseInt(data['Time car 1'], 10),
-      'Time car 2': parseInt(data['Time car 2'], 10),
-      'Location car 1': parseInt(data['Location car 1'], 10),
-      'Location car 2': parseInt(data['Location car 2'], 10)
+      'Time car 1': parseFloat(data['Time car 1'], 10),
+      'Time car 2': parseFloat(data['Time car 2'], 10),
+      'Location car 1': parseFloat(data['Location car 1'], 10),
+      'Location car 2': parseFloat(data['Location car 2'], 10)
     }))
   }
